add 英雄抽奖 英雄属性修改为固定值，增加不同星级 随机范围成长率     英雄升星移除属性增加，战力增加，矿工增加，增加返回英雄详细信息。     增加英雄升级功能

+++ gm-domain/src/main/java/com/gm/modules/user/entity/UserEntity.java
@@ -61,7 +61,7 @@ public class UserEntity implements Serializable {
 	/**
 	 * 用户级别
 	 */
-	private Long userLevelId;
+	private Integer userLevelId;
 	/**
 	 * 累计获得的经验
 	 */
